+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+### Removed
+
+- '===' after battery name.
+
 ### Fixed
 
 - If 59 minutes where rounded to 60, duration showed as "X hours, 60 minutes" instead of "X+1 hours, 0 minutes".
+- Inconsistent hour/minutes output formatting.
 
 ## [0.1.3] - 2022-08-23